refactor: extract CalibreServerManager from ModelData
