+++ .agents/plans/REFACTOR_PLAN.claude.md
